backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic-settings==2.7.1
openai==1.61.0
python-multipart==0.0.20
